260b45d Merge branch 'release'
293d1af Increment version to VTK 5.10.1
bf3ac65 Merge branch 'release'
d99c409 Merge branch 'netcdf_mangling_fixes_rel' into release
b895324 Merge topic 'netcdf_mangling_release_merge' into master
ed2afac Merge topic 'FixTestEmptyInput' into master
4977914 Merge topic 'CubeAxesTransparency' into master
2ffb6a7 Merge topic 'chartxyz-scene-bug' into master
1b40557 Merge topic 'fix-glsl-shader-device-adaptor2' into master
33dae29 Merge topic 'vtk_warning_fixes' into master
da39820 Merge topic 'GL2PSSyncUpstream' into master
8de5a0f really suppress linker warning
d76ba07 Merge topic 'fix-amr-windows-warnings' into master
b9a220c Merge topic 'Coverage_TimePointUtility' into master
13ddc51 Sync with svn version of gl2ps.
751913d Implement HasTranslucentPolygonalGeometry for AxisActors.
...

+++ Common/ExecutionModel/Testing/Cxx/TestTemporalSupport.cxx
@@ -130,7 +130,8 @@ public:
     vtkInformation *outInfo = outputVector->GetInformationObject(0);
     double range[2]= {0,9};
     outInfo->Set(vtkStreamingDemandDrivenPipeline::TIME_RANGE(), range,2);
-    outInfo->Set(vtkStreamingDemandDrivenPipeline::TIME_STEPS(), &TimeSteps[0], TimeSteps.size());
+    outInfo->Set(vtkStreamingDemandDrivenPipeline::TIME_STEPS(),
+                 &TimeSteps[0], static_cast<int>(TimeSteps.size()));
     if(this->HasTimeDependentData)
       {
       outInfo->Set(vtkStreamingDemandDrivenPipeline::TIME_DEPENDENT_INFORMATION(),1);
